Add interfaces and implementations for Nats FW Win
Add code to the agent to secure the remote mbus that
bosh provides. Provide a mechanism to restrict workloads
on the VM other than the agent itself to connect to BOSHs
mbus.

On VM creation bosh provides the Agent on the VM with creds
to connect to Nats. Worklaoads that live on the vm could
extract these credentials and use them to connect to nats and
extract further credentials that are being sent over nats once
the rendering of jobs takes place.

This creates a FW rule within the agent bootstrap to limit which
binaries can create outgoing requests to the nats endpoint.

The created FW rules are ephemeral and exist as long as the agent
process runs. This avoids the necessity for cleaning up FW rules if
the remote address of the nats should change. If a process starts
before the rule exists, the rule will still apply once it exists.

Updated tests relying on SetupNetwork calls with new method signature